submission

+++ api/controller/taskController.js
@@ -2,7 +2,8 @@ const {CreateTask,FindTaskUserById,UpdateTask,DeleteTask,GetTaskbyId,SearchTask,
 
 const AddTask = async (req , res) => {
     try {
-        const { _id ,title, description,status,dueDate,recurrence} = req.body;
+        const { _id } = req.user;
+        const {title, description,status,dueDate,recurrence} = req.body;
         const newTask = await CreateTask({_id ,title, description,status,dueDate,recurrence});
         return res.status(201).json({ message: 'Task added successfully',data: newTask});
     } catch (error) {
@@ -15,7 +16,6 @@ const GetAllTaskbyUser = async (req , res) => {
         const alltask = await FindTaskUserById(req);
         return res.status(201).json({'data': alltask});
     } catch (error) {
-        console.log("🚀 ~ file: taskController.js:18 ~ GetAllTaskbyUser ~ error:", error)
         res.status(400).json({ message: 'Bad Request' })
     }
 }
@@ -24,16 +24,26 @@ const GetTaskId = async (req , res) => {
     try {
         const taskId = req.params.id;
         const oldTask = await GetTaskbyId(taskId)
-        return res.status(200).json({data: oldTask });
+        if (oldTask){
+            return res.status(200).json({data: oldTask });
+        }else{
+            return res.status(404).json({message:'Task not Found'});
+        }
     } catch (error) {
         res.status(400).json({ message: 'Bad Request' })
     }
 }
 const UpdateTaskbyId = async (req , res) => {
     try {
-        const updatedTask = await UpdateTask(req)
-        return res.status(200).json({ message: 'Task updated successfully'});
+        const existingtask = await GetTaskbyId(req?.params?.id)
+        if (existingtask){
+            const updatedTask = await UpdateTask(req)
+            return res.status(200).json({ message: 'Task updated successfully'});
+        }else{
+            return res.status(404).json({ message: 'Task Not Found'});
+        }
     } catch (error) {
+        console.log("🚀 ~ file: taskController.js:45 ~ UpdateTaskbyId ~ error:", error)
         res.status(400).json({ message: 'Bad Request' })
     }
 }
@@ -41,11 +51,15 @@ const UpdateTaskbyId = async (req , res) => {
 const DeleteTaskbyId = async (req , res) => {
     try {
         const taskId = req.params.id;
-        await DeleteTask(taskId)
-        await DeleteUserTask(req)
-        return res.status(200).json({ message: 'Task deleted successfully'});
+        const deletedtask = await DeleteTask(taskId)
+        if (deletedtask){
+            await DeleteUserTask(req)
+            return res.status(200).json({ message: 'Task deleted successfully'});
+        }else{
+            return res.status(404).json({ message: 'Task Not Found'});
+        }
     } catch (error) {
-        console.log("🚀 ~ file: taskController.js:48 ~ DeleteTaskbyId ~ error:", error)
+        console.log("🚀 ~ file: taskController.js:60 ~ DeleteTaskbyId ~ error:", error)
         res.status(400).json({ message: 'Bad Request' })
     }
 }
